third_party/blink/web_tests/external/wpt/editing/run/justifyright_1-1000-expected.txt

This is a testharness.js-based test.
Found 125 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[]bar<p>extra" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobar</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobar</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:right\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:right\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo</span>{}<span>bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:right\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<div style=\\"text-align:right\\">foobar</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><span>foo</span><span>bar</span></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<span>foo[</span><span>]bar</span><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobarbaz</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobarbaz</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobar<b>bazqoz</b>quz</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar<b>bazqoz</b>quz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foobar<b>bazqoz</b>quz</div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar<b>bazqoz</b>quz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "foo[bar<b>baz]qoz</b>quz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobar</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobar</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[]bar<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><p>foobarbaz</p></div><p>extra</p>" but got "<p style=\\"text-align:right\\">foobarbaz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<p>foo[bar]baz<p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:right\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:right\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:right\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><h1>foobarbaz</h1></div><p>extra</p>" but got "<h1 style=\\"text-align:right\\">foobarbaz</h1><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<h1>foo[bar]baz</h1><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:right\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:right\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:right\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><pre>foobarbaz</pre></div><p>extra</p>" but got "<pre style=\\"text-align:right\\">foobarbaz</pre><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<pre>foo[bar]baz</pre><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:right\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:right\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:right\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><xmp>foobarbaz</xmp></div><p>extra</p>" but got "<xmp style=\\"text-align:right\\">foobarbaz</xmp><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<xmp>foo[bar]baz</xmp><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><div style=\\"text-align:right\\"><p>foo</p></div><p>bar</p></center><p>extra</p>" but got "<center><p style=\\"text-align:right\\">foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><div style=\\"text-align:right\\"><p>foo</p></div><p>bar</p></center><p>extra</p>" but got "<center><p style=\\"text-align:right\\">foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><div style=\\"text-align:right\\"><p>foo</p></div><p>bar</p></center><p>extra</p>" but got "<center><p style=\\"text-align:right\\">foo</p><p>bar</p></center><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "center" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyright",""]] "<center><p>[foo]<p>bar</center><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<center><div style=\\"text-align:right\\"><p>foo</p></div><p>bar</p></center><p>extra</p>" but got "<center><p style=\\"text-align:right\\">foo</p><p>bar</p></center><p>extra</p>"
Harness: the test ran to completion.
